+++ examples/src/test/scala/org/datatools/bigdatatypes/formats/CustomFormatsSpec.scala
@@ -40,7 +40,7 @@ class CustomFormatsSpec extends UnitSpec {
             "myDummy",
             StandardSQLTypeName.STRUCT,
             FieldList.of(
-              fields: _*
+              fields*
             )
           )
           .setMode(Mode.REQUIRED)