+++ spark/src/test/scala/org/datatools/bigdatatypes/spark/SparkTypeConversionSpec.scala
@@ -1,14 +1,14 @@
 package org.datatools.bigdatatypes.spark
 
 import org.apache.spark.sql.types.{IntegerType, StringType, StructField, StructType}
-import org.datatools.bigdatatypes.TestTypes._
+import org.datatools.bigdatatypes.TestTypes.*
 import org.datatools.bigdatatypes.UnitSpec
-import org.datatools.bigdatatypes.basictypes._
-import org.datatools.bigdatatypes.basictypes.SqlType._
-import org.datatools.bigdatatypes.basictypes.SqlTypeMode._
+import org.datatools.bigdatatypes.basictypes.*
+import org.datatools.bigdatatypes.basictypes.SqlType.*
+import org.datatools.bigdatatypes.basictypes.SqlTypeMode.*
 import org.datatools.bigdatatypes.conversions.{SqlInstanceConversion, SqlTypeConversion}
 import org.datatools.bigdatatypes.formats.Formats.implicitDefaultFormats
-import org.datatools.bigdatatypes.spark.SparkTypeConversion._
+import org.datatools.bigdatatypes.spark.SparkTypeConversion.*
 
 /** Reverse conversion, from Spark types to [[SqlType]]s
   */
@@ -20,13 +20,13 @@ class SparkTypeConversionSpec extends UnitSpec {
 
   "StructField nullable" should "be converted into Nullable SqlType" in {
     val sf = StructField("myInt", IntegerType, nullable = true)
-    sf.getType shouldBe SqlInt(Nullable)
+    sf.asSqlType shouldBe SqlInt(Nullable)
     SqlInstanceConversion[StructField].getType(sf) shouldBe SqlInt(Nullable)
   }
 
   "StructField required" should "be converted into Required SqlType" in {
     val sf = StructField("myInt", IntegerType, nullable = false)
-    sf.getType shouldBe SqlInt()
+    sf.asSqlType shouldBe SqlInt()
     SqlInstanceConversion[StructField].getType(sf) shouldBe SqlInt(Required)
   }
 
@@ -36,48 +36,48 @@ class SparkTypeConversionSpec extends UnitSpec {
     val st = StructType(List(sf, sf2))
     val expected = SqlStruct(List(("myInt", SqlInt()), ("myString", SqlString(Nullable))))
     SqlInstanceConversion[StructType].getType(st) shouldBe expected
-    st.getType shouldBe expected
+    st.asSqlType shouldBe expected
   }
 
   behavior of "Common Test Types for Spark"
 
   "Spark Schema with Option" should "be converted into SqlTypes with nullable" in {
-    val sqlType: SqlType = SqlTypeToSpark[BasicOption].sparkSchema.getType
+    val sqlType: SqlType = SqlTypeToSpark[BasicOption].sparkSchema.asSqlType
     sqlType shouldBe basicOption
   }
 
   "basic Spark Schema" should "be converted into SqlTypes" in {
-    val sqlType: SqlType = SqlTypeToSpark[BasicTypes].sparkSchema.getType
+    val sqlType: SqlType = SqlTypeToSpark[BasicTypes].sparkSchema.asSqlType
     sqlType shouldBe basicTypes
   }
 
   "Spark Schema with basic options types" should "be converted into nullable SqlTypes" in {
-    val sqlType: SqlType = SqlTypeToSpark[BasicOptionTypes].sparkSchema.getType
+    val sqlType: SqlType = SqlTypeToSpark[BasicOptionTypes].sparkSchema.asSqlType
     sqlType shouldBe basicOptionTypes
   }
 
   "Spark Schema with List" should "be converted into Repeated type" in {
-    val sqlType: SqlType = SqlTypeToSpark[BasicList].sparkSchema.getType
+    val sqlType: SqlType = SqlTypeToSpark[BasicList].sparkSchema.asSqlType
     sqlType shouldBe basicWithList
   }
 
   "Spark Schema with nested object" should "be converted into SqlTypes" in {
-    val sqlType: SqlType = SqlTypeToSpark[BasicStruct].sparkSchema.getType
+    val sqlType: SqlType = SqlTypeToSpark[BasicStruct].sparkSchema.asSqlType
     sqlType shouldBe basicNested
   }
 
   "Spark Schema with optional nested object" should "be converted into SqlTypes" in {
-    val sqlType: SqlType = SqlTypeToSpark[BasicOptionalStruct].sparkSchema.getType
+    val sqlType: SqlType = SqlTypeToSpark[BasicOptionalStruct].sparkSchema.asSqlType
     sqlType shouldBe basicOptionalNested
   }
 
   "Spark Schema with Struct List" should "be converted into Repeated Struct type" in {
-    val sqlType: SqlType = SqlTypeToSpark[ListOfStruct].sparkSchema.getType
+    val sqlType: SqlType = SqlTypeToSpark[ListOfStruct].sparkSchema.asSqlType
     sqlType shouldBe basicNestedWithList
   }
 
   "Spark Schema with extended types" should "be converted into Struct with extended types" in {
-    val sqlType: SqlType = SqlTypeToSpark[ExtendedTypes].sparkSchema.getType
+    val sqlType: SqlType = SqlTypeToSpark[ExtendedTypes].sparkSchema.asSqlType
     sqlType shouldBe extendedTypes
   }